requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asn1crypto==0.24.0        # via cryptography
astroid==2.2.5            # via conan, pylint
bottle==0.12.16           # via conan
certifi==2019.3.9         # via requests
cffi==1.12.3              # via cryptography
chardet==3.0.4            # via requests
cmake==3.14.4
colorama==0.4.1           # via conan
conan-package-tools==0.27.0
conan==1.15.2
cryptography==2.3.1
deprecation==2.0.6        # via conan
distro==1.1.0             # via conan
fasteners==0.15           # via conan
future==0.16.0            # via conan
idna==2.6
isort==4.3.20             # via pylint
jinja2==2.10.1            # via conan
lazy-object-proxy==1.4.1  # via astroid
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via pylint
monotonic==1.5            # via fasteners
ninja==1.9.0.post1
node-semver==0.6.1        # via conan
packaging==19.0           # via deprecation
patch==1.16               # via conan
pluginbase==0.7           # via conan
pycparser==2.19           # via cffi
pygments==2.4.2           # via conan
pyjwt==1.7.1              # via conan
pylint==2.3.1             # via conan
pyopenssl==18.0.0
pyparsing==2.4.0          # via packaging
pyyaml==5.1               # via conan
requests==2.22.0          # via conan
six==1.12.0               # via astroid, conan, conan-package-tools, cryptography, fasteners, packaging, pyopenssl
tabulate==0.8.2           # via conan-package-tools
tqdm==4.32.1              # via conan
typed-ast==1.3.5          # via astroid
urllib3==1.25.3           # via requests
wrapt==1.11.1             # via astroid
